:bug:  Fix: 요약, 키워드, 리마인드 퀴즈 수정 및 재생성 동시 수정 제어
- 수정중인사람은 컬럼을 USING을 사용하여 다음 사용자가 수정을 못하도록 막음
- 읽는 것도 막음


Former-commit-id: 2e01385a64f9392b4d7d1647ee748bb3f4d90aed

+++ Backend/src/main/java/com/baekopa/backend/domain/meeting/service/MeetingService.java
@@ -27,6 +27,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
+import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -117,8 +119,23 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingSummaryResponseDTO createSummary(Long meetingId) {
+    public synchronized MeetingSummaryResponseDTO createSummary(Long meetingId) {
         String summaryUrl = fastUrl + "/studies/summary";
+
+        Meeting meeting = meetingRepository.findById(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
+        if (meetingSummaryRepository.existsByMeetingAndDeletedAtIsNull(meeting)) {//기존의 미팅 sumary 데이터가 있을 경우
+            MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
+                    .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+            if (meetingSummary.getStatus() == IsolationEnum.USING) {//이미 누군가 meetingSummary값을 수정 중이라면
+                throw new BusinessException(ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS, ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS.getMessage());
+            }
+            //기존의 미팅요약을 삭제
+            meetingSummaryRepository.delete(meetingSummary);
+            meetingSummaryRepository.flush();
+        }
+        //새로운 미팅 생성(사용중인값으로 생성)
+        MeetingSummary meetingSummary = meetingSummaryRepository.saveAndFlush(MeetingSummary.of(meeting));
 
         String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findById(meetingId)
                         .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage())))
@@ -143,23 +160,25 @@ public class MeetingService {
         HttpEntity<Object> requestEntity = new HttpEntity<>(jsonText, headers);
 
         // db 저장
-        Meeting meeting = meetingRepository.findById(meetingId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         MeetingSummaryDTO meetingSummaryDTO = restTemplate.postForObject(summaryUrl, requestEntity, MeetingSummaryDTO.class);
-        MeetingSummary meetingSummary = MeetingSummary.of(meeting, meetingSummaryDTO.getSummaryText());
-        meetingSummaryRepository.save(meetingSummary);
+        meetingSummary.updateMeetingSummary(meetingSummaryDTO.getSummaryText());
         MeetingSummaryResponseDTO meetingSummaryResponseDTO = MeetingSummaryResponseDTO.from(meetingSummaryDTO);
 
         return meetingSummaryResponseDTO;
     }
 
     public MeetingSummaryResponseDTO getMeetingSummary(Long meetingId) {
-        MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId).orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        if(meetingSummary.getStatus()==IsolationEnum.USING){
+            throw new BusinessException(ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS, ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS.getMessage());
+        }
+
         return MeetingSummaryResponseDTO.getMeetingSummary(meetingSummary);
     }
 
     @Transactional
-    public MeetingSummaryResponseDTO updateCreateSummary(Long meetingId) {
+    public synchronized MeetingSummaryResponseDTO updateCreateSummary(Long meetingId) {
         String summaryUrl = fastUrl + "/studies/summary";
 
         String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findById(meetingId)
@@ -193,20 +212,48 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingSummaryResponseDTO updateMeetingSummary(MeetingSummaryUpdateDTO meetingSummaryUpdateDTO, Long meetingId) {
+    public synchronized MeetingSummaryResponseDTO updateMeetingSummary(MeetingSummaryUpdateDTO meetingSummaryUpdateDTO, Long meetingId) {
         MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        meetingSummary.setUpdateStatus(IsolationEnum.USING);
+        meetingSummaryRepository.flush();
+
         meetingSummary.updateMeetingSummary(meetingSummaryUpdateDTO.getSummaryText());
 
         return MeetingSummaryResponseDTO.getMeetingSummary(meetingSummary);
     }
 
     @Transactional
-    public MeetingRemindQuizResponseDTO createMeetingRemindQuiz(Long meetingId) {
+    public synchronized MeetingRemindQuizResponseDTO createMeetingRemindQuiz(Long meetingId) {
         String remindQuizUrl = fastUrl + "/studies/quiz";
+        LocalDateTime openDate = null;
+
+        Meeting meeting = meetingRepository.findById(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
 
         MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        // todo 읽기 권한 체크 더 하기
+        if(meetingSummary.getStatus()==IsolationEnum.USING){
+            throw new BusinessException(ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS, ErrorCode.MEETING_REMIND_QUIZ_NOT_FOUND.getMessage());
+        }
+
+        if (remindQuizRepository.existsByMeetingAndDeletedAtIsNull(meeting)) { // 리마인드 퀴즈가 이미 존재한다면..?
+            RemindQuiz remindQuiz = remindQuizRepository.findByMeetingAndDeletedAtIsNull(meeting)
+                    .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_REMIND_QUIZ_NOT_FOUND, ErrorCode.MEETING_REMIND_QUIZ_NOT_FOUND.getMessage()));
+            openDate = remindQuiz.getOpenDate();
+
+            if (remindQuiz.getStatus() == IsolationEnum.USING) {
+                throw new BusinessException(ErrorCode.REMIND_QUIZ_CAN_CNT_ACCESS, ErrorCode.MEETING_REMIND_QUIZ_NOT_FOUND.getMessage());
+            }
+            remindQuizRepository.delete(remindQuiz);
+            remindQuizRepository.flush();
+        }
+        if(openDate==null){
+            openDate=LocalDateTime.now(ZoneId.of("Asia/Seoul")).plusDays(7);
+        }
+        RemindQuiz remindQuiz=RemindQuiz.from(meeting, openDate);
+        remindQuizRepository.saveAndFlush(remindQuiz);
 
         CreateMeetingRemindQuizDTO createMeetingRemindQuizDTO = CreateMeetingRemindQuizDTO.from(meetingSummary);
         ObjectMapper objectMapper = new ObjectMapper();
@@ -225,11 +272,10 @@ public class MeetingService {
 
         MeetingRemindQuizResponseDTO meetingRemindQuizResponseDTO = restTemplate.postForObject(remindQuizUrl, requestEntity, MeetingRemindQuizResponseDTO.class);
 
-        Meeting meeting = meetingRepository.findById(meetingId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         // db 저장
         RemindQuizDTO remindQuizDTO = RemindQuizDTO.of(meetingRemindQuizResponseDTO.getQuiz());
-        remindQuizRepository.save(RemindQuiz.from(meeting, remindQuizDTO));
+        //remindQuizRepository.save(RemindQuiz.from(meeting, remindQuizDTO));
+        remindQuiz.updateRemindQuiz(remindQuizDTO.getQuiz());
 
         return meetingRemindQuizResponseDTO;
     }
@@ -269,23 +315,27 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingKeywordListDTO createMeetingKeyword(Long meetingId) {
+    public synchronized MeetingKeywordListDTO createMeetingKeyword(Long meetingId) {
         Meeting meeting = meetingRepository.findById(meetingId).orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         //이미 기존의 키워드가 존재 한다면?
-        if (meetingKeywordRepository.existsByMeetingAndDeletedAtIsNull(meeting)) {
-            List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
-
-            for (MeetingKeyword meetingKeyword : meetingKeywordList) {
-                meetingKeywordRepository.delete(meetingKeyword);
-            }
-        }
-
-        String keywordUrl = fastUrl + "/studies/keyword";
 
         String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findById(meetingId)
                         .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage())))
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SCRIPT_NOT_FOUND, ErrorCode.MEETING_SCRIPT_NOT_FOUND.getMessage()))
                 .getScriptContent();
+
+        if (meetingKeywordRepository.existsByMeetingAndDeletedAtIsNull(meeting)) {
+            List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
+
+            for (MeetingKeyword meetingKeyword : meetingKeywordList) {
+                if(meetingKeyword.getStatus()==IsolationEnum.USING)
+                    throw new BusinessException(ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS, ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS.getMessage());
+                meetingKeywordRepository.delete(meetingKeyword);
+            }
+            meetingKeywordRepository.flush();
+        }
+
+        String keywordUrl = fastUrl + "/studies/keyword";
 
         //originalText = originalText.replace(".", ".\n");
 
@@ -310,7 +360,7 @@ public class MeetingService {
         // db 저장
 
         for (int i = 0; i < MeetingKeywordResponseDTO.getKeyword().size(); i++) {
-            meetingKeywordRepository.save(MeetingKeyword.of(meeting, MeetingKeywordResponseDTO.getKeyword().get(i)));
+            meetingKeywordRepository.saveAndFlush(MeetingKeyword.of(meeting, MeetingKeywordResponseDTO.getKeyword().get(i)));
         }
 
         List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
@@ -330,6 +380,9 @@ public class MeetingService {
         List<MeetingKeywordDTO> meetingKeywordDTOList = new ArrayList<>();
 
         for (MeetingKeyword meetingKeyword : meetingKeywordList) {
+            if(meetingKeyword.getStatus()==IsolationEnum.USING){
+                throw new BusinessException(ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS, ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS.getMessage());
+            }
             MeetingKeywordDTO meetingKeywordDTO = MeetingKeywordDTO.from(meetingKeyword);
             meetingKeywordDTOList.add(meetingKeywordDTO);
         }
@@ -337,11 +390,10 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingKeywordListDTO updateMeetingKeyword(Long meetingId, UpdateMeetingKeywordListDTO updateMeetingKeywordListDTO) {
+    public synchronized MeetingKeywordListDTO updateMeetingKeyword(Long meetingId, UpdateMeetingKeywordListDTO updateMeetingKeywordListDTO) {
         for (UpdateMeetingKeywordDTO updateMeetingKeywordDTO : updateMeetingKeywordListDTO.getUpdateMeetingKeywordList()) {
             MeetingKeyword meetingKeyword = meetingKeywordRepository.findByIdAndDeletedAtIsNull(updateMeetingKeywordDTO.getGroupKeywordId())
                     .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_KEYWORD_NOT_FOUND, ErrorCode.MEETING_KEYWORD_NOT_FOUND.getMessage()));
-
             meetingKeyword.updateMeetingKeyword(updateMeetingKeywordDTO.getKeyword());
         }
